Crate srml_democracy

Democracy Module

The Democracy module handles administration of general stakeholder voting.

democracy::Trait

Call

Module

Overview

Terminology

Proposal: A proposal can take several forms. In a spending proposal, a proposer proposes to send funds to a beneficiary account (which may be a normal account or a smart contract). A proposal may include set_code, which would upgrade the runtime. The proposer must attach a bond with the proposal, which will be taken if it is rejected.

Referendum: A mechanism for voting on proposals. A referendum includes a single proposal, a block number when voting will end, a threshold mechanism, and a delay (in blocks) to wait before deploying the associated proposal (if successfully passed). Referenda can be started in three ways: from the public, from a unanimous Council vote, or a majority Council vote. Each method of starting a referendum comes with a different default vote threshold mechanism.

Vote threshold mechanisms: Different criteria for passing or rejecting a referendum (e.g. supermajority for, supermajority against, simple majority). Supermajority thresholds are based on Adaptive Quorum Biasing such that the required supermajority increases with lower turnout. As turnout approaches 100%, the required majority approaches 50%.

Table of Referenda: A set of referenda that are currently open for voting.

Proposal Terminology

Validity: The proposer's deposit must be above the minimum required amount and must have sufficient account balance to transfer it into a reserve.

Sponsorship: Sponsor (second) a public proposal using the second call that anyone may execute by signing and submitting an extrinsic.

Sponsorship validity: For a sponsorship to be valid, the sponsored proposal must exist and a deposit from its proposer must have been reserved. The sponsor must have sufficient account balance to reserve a matching deposit.

Elevation process: If checking the configured LaunchPeriod indicates that a new public referendum should be launched, then the public proposal index with the largest locked deposit amount is declared the winning proposal. This proposal is removed from PublicProps and becomes a public referendum. The accounts that locked a deposit into this winning proposal are refunded their reserved deposit.

Referenda Terminology

Start: Start a public referendum using the start_referendum call that anyone may execute by signing and submitting an extrinsic. It is allocated the next referendum index, which is mapped to its corresponding voting period expiry block, the proposal it relates to, and the given voting threshold.

Cancellation: Remove all information about a referendum.

Validity: The new referendum must not have a voting period that ends before any existing referenda.

Voting: Voters may vote on a public referendum using the vote call that anyone may execute by signing and submitting an extrinsic. Voters and their votes (yay or nay) for a public referendum are stored in VotersFor and VoteOf, respectively.

Vote delegation: Voters may delegate and undelegate their votes for an amount of lock periods.

Proxy account: Stash accounts (see the Staking module) may add or remove a proxy account to vote on a public referendum on behalf of the stash account.

Vote validity: The referendum being voted on must be an active referendum index of the ReferendumInfoOf mapping. The voter (transactor) must have a balance above zero to signal approval.

Maturity: A mature referendum is one that expires at the current block.

Vote tallying, passing, and execution: When a block is finalized, the Democracy module will search for mature referenda and tally their votes, then remove them from the Table of Referenda. If a referendum's vote tally meets its vote threshold, then it will be passed and executed. Lastly, NextTally is incremented to determine the next public referendum index to tally.

Goals

The Democracy module in Substrate is designed to make the following possible:

Create and sponsor public proposals.

Elevate public proposals to the Table of Referenda.

Start and cancel public referenda.

Vote on public referenda.

Delegate proxy voting rights of public referenda.

Tally votes of public referenda.

Pass and execute maturing public referenda.

Interface

Dispatchable Functions

propose - Propose an action to be taken.

second - Second (sponsor) a proposal.

vote - Vote on a referendum.

proxy_vote - Vote on a referendum using a proxy account on behalf of a stash account.

start_referendum - Start a referendum.

cancel_referendum - Remove a referendum.

cancel_queued - Cancel a proposal queued for enactment.

on_finalize - Called when a block is finalized.

set_proxy - Specify a proxy account. Called by the stash account.

resign_proxy - Clear the proxy account. Called by the proxy account.

remove_proxy - Clear the proxy account. Called by the stash account.

delegate - Delegate vote.

undelegate - Undelegate vote.

Public Functions

locked_for - Get the balance locked in support of a proposal.

is_active_referendum - Return true if the given referendum index corresponds to an ongoing referendum.

active_referendums - Get all referenda that are currently active and their corresponding info. Equivalent to the Table of Referenda.

maturing_referendums_at - Get all referenda ready for tally at block n.

tally - Tally the votes for the current proposal.

force_proxy - Forcibly insert a proxy voter for a stash account.

internal_start_referendum - Start a referendum. Can be called directly by the council.

internal_cancel_referendum - Remove a referendum. Can be called directly by the council.

Usage

Prerequisites

Import the Democracy module and types and derive your runtime's configuration traits from the Democracy module trait.

Example from the SRML

The Council module uses the Democracy module for voting.

use srml_support::{decl_module, dispatch::Result};
pub trait Trait: democracy::Trait { }

decl_module! {
	pub struct Module<T: Trait> for enum Call where origin: T::Origin {

		fn proxy_set_approvals(origin, votes: Vec<bool>, index: VoteIndex) -> Result {
			let who = <democracy::Module<T>>::proxy(ensure_signed(origin)?).ok_or("not a proxy")?;
			// set approvals
			Ok(())
		}
	}
}

Genesis Config

The Democracy module depends on the GenesisConfig.

Related Modules

Staking

Council

Structs

Delegations
Get the account (and lock periods) to which another account is delegating vote.

DepositOf
Those who have locked a deposit.

DispatchQueue
Queue of successful referenda to be dispatched.

GenesisConfig

LaunchPeriod
How often (in blocks) new public referenda are launched.

MaxLockPeriods
The maximum number of additional lock periods a voter may offer to strengthen their vote.

MinimumDeposit
The minimum amount to be used as a deposit for a public referendum proposal.

Module

NextTally
The next referendum index that should be tallied.

Proxy
Who is able to vote for whom. Value is the fund-holding account, key is the vote-transaction-sending account.

PublicDelay
The delay before enactment for all public referenda.

PublicPropCount
The number of (public) proposals that have been made so far.

PublicProps
The public proposals. Unsorted. T::AccountId refers to the account that proposed.

ReferendumCount
The next free referendum index, aka the number of referenda started so far.

ReferendumInfo
Info regarding an ongoing referendum.

ReferendumInfoOf
Information concerning any given referendum.

Vote
A number of lock periods, plus a vote, one way or the other.

VoteOf
Get the vote in a given referendum of a particular voter. The result is meaningful only if voters_for includes the voter when called with the referendum (you'll get the default Vote value otherwise). If you don't want to check voters_for, then you can also check for simple existence with VoteOf::exists first.

VotersFor
Get the voters for the current proposal.

VotingPeriod
How often (in blocks) to check for new votes.

Enums

Call

RawEvent
Events for this module.

VoteThreshold
A means of determining if a vote is above the required threshold to pass.

Traits

Approved

Trait

Type Definitions

Event
RawEvent specialized for the configuration Trait

LockPeriods
A number of lock periods.

PropIndex
A proposal index.

ReferendumIndex
A referendum index.